requirements.txt
torch
torchvision
numpy
PyYAML
scikit-learn
Pillow
pytest